% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_geogwilen.R
\name{charge_shp_geogwilen}
\alias{charge_shp_geogwilen}
\title{charge_shp_geogwilen}
\usage{
charge_shp_geogwilen(shp, scheme, shp_emprise = NULL, col_geometry = NULL)
}
\arguments{
\item{shp}{Name of the geogwilen table}

\item{scheme}{Name of the scheme were geogwilen table is stocked}

\item{shp_emprise}{sf object defining the extent of the query (optional)}

\item{col_geometry}{geometry name in the data}
}
\value{
sf object with result of the query
}
\description{
charge_shp_geogwilen
}
\examples{

config <- yaml::read_yaml("C://workspace//gwilenalim//yaml//config.yml")
# chargement de l'objet shp_Vilaine
datafile <- system.file("shp_Vilaine.rdata", package = "geogwilenR")
load(file=datafile)

# chargement du bv de la Vilaine
bv<-charge_shp_geogwilen(shp="bassin_vilaine", 
                     scheme="r300_territoire_sage")


# test de la fonction
pop_carroyee<-charge_shp_geogwilen(shp="filosofi2015_carreaux_200m_metropole", 
                     scheme="r350_eco_socio", 
                     shp_emprise=bv) 

pop_carroyee<-charge_shp_geogwilen(shp="Filosofi2017_carreaux_200m_bzh_pdl", 
                     scheme="r350_eco_socio", 
                     shp_emprise=bv) 

# test de la fonction
step<-charge_shp_geogwilen(shp="r621_step_vilaine_xy_ste", 
                     scheme="r621_step", 
                     shp_emprise=NULL) 


step_vilaine<-charge_shp_geogwilen(shp="bdd_step_fr", 
                     scheme="m621_step", 
                     shp_emprise=bv) 


# 
# options(viewer = NULL) 
# library(mapview)
# mapview(shp_Vilaine) + mapview(step_vilaine)
# 


}
