% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogwilenR-package.R
\docType{package}
\name{geogwilenR-package}
\alias{geogwilenR}
\alias{geogwilenR-package}
\title{geogwilenR: Package Pour Interroger La Base Geogwilen}
\description{
Ce package permet d'importer des données depuis la base geogwilen sous R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnthonyDEBUR/geogwilenR}
  \item Report bugs at \url{https://github.com/AnthonyDEBUR/geogwilenR/issues}
}

}
\author{
\strong{Maintainer}: Anthony DE BURGHRAVE \email{anthony.deburghrave@eaux-et-vilaine.bzh}

}
\keyword{internal}
